fggd

+++ Presentation/EIRS.Web/GISModels/ApprovalFlow.cs
@@ -10,9 +10,9 @@ namespace EIRS.Web.GISModels
         public string Rin { get; set; }
         public string AssessmentRefNo { get; set; }
         public string TaxPayerName { get; set; }
-        public decimal Amount { get; set; }
+        public double Amount { get; set; }
         public string Status { get; set; }
         public long AssessmentId { get; set; }
-        public int? Id { get; set; }
+        public long? Id { get; set; }
     }
 }
